Removed Yum Update
Is there any particular reason an OS update is needed? This is less than ideal.

If this command was inteded to download the package lists from the repositories and "update" them, YUM already does this by default on Yum install. 

It does not require to "update" as APT does.

"yum update" will update all the packages to the latest version... 

yum update package_name
    Used to update the specified packages to the latest available version. If no packages are specified, then yum will attempt to update all installed packages. 

https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/security_guide/chap-keeping_your_system_up-to-date