"Scale Out" not "Scale Up" for Vertical Scaling
In AWS, the phrase “Scale UP the EC2 instance” means that you have to UPGRADE (horizontally scale) the instance to a higher instance type ( e.g. t3 small to t3 large). 

When you say, “Scale OUT your EC2 instances" -- it means that you have to increase (vertically scale ) the number of EC2 instances using Auto Scaling.

On this page, there is an ambiguity in the terms used. It's using "scale UP" for a scaling operation. It should be "scale OUT" the number of EC2 instances.

All of the other AWS pages below use the correct phrase: "scale out" : 
 
https://docs.aws.amazon.com/autoscaling/ec2/userguide/asg-capacity-limits.html
https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-scaling-simple-step.html
https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/create_deploy_NET.managing.as.html


The same format is also used in Azure documentation. I understand that "Scale UP" and "Scale OUT" seems to be synonymous but I think we should to use the most accurate terms to avoid confusion.  

Cheers,
Jon Bonso

https://www.linkedin.com/in/jonbonso/